+++ packages/core/types/src/http/workflow-execution/admin/queries.ts
@@ -1,3 +1,4 @@
+import { OperatorMap } from "../../.."
 import { FindParams } from "../../common"
 
 export interface AdminGetWorkflowExecutionsParams extends FindParams {
@@ -13,4 +14,12 @@ export interface AdminGetWorkflowExecutionsParams extends FindParams {
    * Filter by the ID of the workflow to retrieve workflow executions for a specific workflow.
    */
   workflow_id?: string | string[]
+  /**
+   * Filter by the state of the workflow execution.
+   */
+  state?: string | string[]
+  /**
+   * Filter by the creation date of the workflow execution.
+   */
+  created_at?: OperatorMap<string>
 }